refactor(frontend): extract v1-free leaf inputs to components/base/inputs (reuse-don't-fork)
