+++ extralit-frontend/components/base/inputs/label-selection/labelSelection.component.spec.js
