+++ extralit-frontend/components/base/inputs/label-selection/useLabelSelectionViewModel.ts
