+++ extralit-frontend/components/base/inputs/ranking/ranking-adapter.test.js
