+++ extralit-frontend/components/base/inputs/ranking/ranking-fakes.js
