+++ extralit-frontend/components/base/inputs/rating/ratingMonoSelection.component.spec.js
